fix: honor scoped mappedToModel overrides in Coder usage-ledger keys (Codex round 26)
normalizeUsageModelKey moved to modelEntries.ts (models.ts cannot import
resolveModelForMetadata without a cycle) and now resolves coder: strings
through resolveModelForMetadata so an explicit scoped mappedToModel wins
before the instance-type-derived identity. Repricing sees only the stored
key and cannot recover a Coder-scoped mapping, so it must be applied at
record time.
