+++ src/common/utils/ai/models.test.ts
@@ -3,7 +3,6 @@ import {
   getExplicitGatewayPrefix,
   normalizeSelectedModel,
   normalizeToCanonical,
-  normalizeUsageModelKey,
   getModelName,
   getAnthropic1MContextMode,
   hasNative1MContext,
@@ -81,54 +80,6 @@ describe("normalizeToCanonical", () => {
 
   it("returns malformed gateway strings unchanged", () => {
     expect(normalizeToCanonical("mux-gateway:no-slash-here")).toBe("mux-gateway:no-slash-here");
-  });
-});
-
-describe("normalizeUsageModelKey", () => {
-  const config = {
-    coder: {
-      apiKeySet: false,
-      isEnabled: true,
-      isConfigured: true,
-      discoveredProviders: [
-        { name: "openai", type: "anthropic" },
-        { name: "prod-anthropic", type: "anthropic" },
-        { name: "llm-proxy", type: "openai-compat" },
-      ],
-    },
-  };
-
-  it("resolves Coder identities to their record-time metadata identity", () => {
-    // TYPE-derived, not name-derived: a cross-typed instance keys under
-    // anthropic:* so later repricing prices the right provider family, and
-    // the key stays resolvable after the instance is removed from
-    // discoveredProviders (the raw gateway key would not).
-    expect(normalizeUsageModelKey("coder:openai/claude-opus-4-5", config)).toBe(
-      "anthropic:claude-opus-4-5"
-    );
-    expect(normalizeUsageModelKey("coder:prod-anthropic/claude-opus-4-5", config)).toBe(
-      "anthropic:claude-opus-4-5"
-    );
-    // Without metadata, the name === type default applies.
-    expect(normalizeUsageModelKey("coder:anthropic/claude-opus-4-5")).toBe(
-      "anthropic:claude-opus-4-5"
-    );
-  });
-
-  it("keeps the raw key for unmappable Coder identities", () => {
-    // openai-compat fronts arbitrary upstreams and unknown instances have no
-    // catalog identity: the raw gateway key is their only durable identity.
-    expect(normalizeUsageModelKey("coder:llm-proxy/llama-3.3-70b", config)).toBe(
-      "coder:llm-proxy/llama-3.3-70b"
-    );
-    expect(normalizeUsageModelKey("coder:mystery/model", config)).toBe("coder:mystery/model");
-  });
-
-  it("canonicalizes non-Coder gateway strings like normalizeToCanonical", () => {
-    expect(normalizeUsageModelKey("mux-gateway:anthropic/claude-opus-4-5")).toBe(
-      "anthropic:claude-opus-4-5"
-    );
-    expect(normalizeUsageModelKey("anthropic:claude-opus-4-5")).toBe("anthropic:claude-opus-4-5");
   });
 });
