+++ src/common/utils/ai/models.ts
@@ -32,32 +32,6 @@ export function isValidModelFormat(model: string): boolean {
 }
 
 /**
- * Usage-ledger key for a model string. Same as normalizeToCanonical, except
- * Coder gateway identities resolve through instance metadata AT RECORD TIME
- * to their stable upstream catalog identity (coder:prod-anthropic/<claude>
- * -> anthropic:<claude>). Two failure modes rule out the alternatives:
- * name-based canonicalization keys a cross-typed instance ({name: "openai",
- * type: "anthropic"}) under openai:<claude>, so later repricing strips the
- * recorded Anthropic costs as unknown OpenAI-model costs; the raw gateway
- * key depends on mutable discoveredProviders metadata, so an authoritative
- * refresh that removes the instance makes the key unresolvable and repricing
- * strips its costs too. The record-time metadata identity survives both, and
- * matches how default-named coder:anthropic/<model> keys were persisted
- * before instance generalization. Unmappable identities (openai-compat
- * instances, unknown instances, a custom provider shadowing the "coder"
- * prefix) keep the raw key — it is their only durable identity.
- */
-export function normalizeUsageModelKey(
-  modelString: string,
-  providersConfig?: Record<string, unknown> | null
-): string {
-  if (modelString.startsWith("coder:")) {
-    return resolveCoderGatewayMetadataModel(modelString, providersConfig) ?? modelString;
-  }
-  return normalizeToCanonical(modelString);
-}
-
-/**
  * Normalize gateway model strings to canonical provider:model format when possible.
  * For gateway-only vendor/model IDs, keep the original gateway-scoped identity.
  */
