+++ src/common/utils/providers/modelEntries.test.ts
@@ -4,6 +4,7 @@ import {
   getModelContextWindowOverride,
   getProviderModelEntryMappedTo,
   normalizeProviderModelEntry,
+  normalizeUsageModelKey,
   resolveModelForMetadata,
 } from "./modelEntries";
 
@@ -108,6 +109,74 @@ describe("resolveModelForMetadata for Coder gateway-scoped models", () => {
     expect(resolveModelForMetadata("coder:anthropic/claude-opus-4-5", config)).toBe(
       "coder:anthropic/claude-opus-4-5"
     );
+  });
+});
+
+describe("normalizeUsageModelKey", () => {
+  const config: ProvidersConfigMap = {
+    coder: {
+      apiKeySet: false,
+      isEnabled: true,
+      isConfigured: true,
+      discoveredProviders: [
+        { name: "openai", type: "anthropic" },
+        { name: "prod-anthropic", type: "anthropic" },
+        { name: "llm-proxy", type: "openai-compat" },
+      ],
+    },
+  };
+
+  test("resolves Coder identities to their record-time metadata identity", () => {
+    // TYPE-derived, not name-derived: a cross-typed instance keys under
+    // anthropic:* so later repricing prices the right provider family, and
+    // the key stays resolvable after the instance is removed from
+    // discoveredProviders (the raw gateway key would not).
+    expect(normalizeUsageModelKey("coder:openai/claude-opus-4-5", config)).toBe(
+      "anthropic:claude-opus-4-5"
+    );
+    expect(normalizeUsageModelKey("coder:prod-anthropic/claude-opus-4-5", config)).toBe(
+      "anthropic:claude-opus-4-5"
+    );
+    // Without metadata, the name === type default applies.
+    expect(normalizeUsageModelKey("coder:anthropic/claude-opus-4-5")).toBe(
+      "anthropic:claude-opus-4-5"
+    );
+  });
+
+  test("honors a scoped mappedToModel override before the type-derived identity", () => {
+    // Record-time pricing resolves mappedToModel via resolveModelForMetadata;
+    // if the KEY discarded the mapping (anthropic:internal-alias), repricing
+    // could not recover the Coder-scoped entry and would zero the costs.
+    const mappedConfig: ProvidersConfigMap = {
+      coder: {
+        apiKeySet: false,
+        isEnabled: true,
+        isConfigured: true,
+        discoveredProviders: [{ name: "prod-anthropic", type: "anthropic" }],
+        models: [
+          { id: "prod-anthropic/internal-alias", mappedToModel: "anthropic:claude-opus-4-5" },
+        ],
+      },
+    };
+    expect(normalizeUsageModelKey("coder:prod-anthropic/internal-alias", mappedConfig)).toBe(
+      "anthropic:claude-opus-4-5"
+    );
+  });
+
+  test("keeps the raw key for unmappable Coder identities", () => {
+    // openai-compat fronts arbitrary upstreams and unknown instances have no
+    // catalog identity: the raw gateway key is their only durable identity.
+    expect(normalizeUsageModelKey("coder:llm-proxy/llama-3.3-70b", config)).toBe(
+      "coder:llm-proxy/llama-3.3-70b"
+    );
+    expect(normalizeUsageModelKey("coder:mystery/model", config)).toBe("coder:mystery/model");
+  });
+
+  test("canonicalizes non-Coder gateway strings like normalizeToCanonical", () => {
+    expect(normalizeUsageModelKey("mux-gateway:anthropic/claude-opus-4-5")).toBe(
+      "anthropic:claude-opus-4-5"
+    );
+    expect(normalizeUsageModelKey("anthropic:claude-opus-4-5")).toBe("anthropic:claude-opus-4-5");
   });
 });
